backend/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorgram.py==1.2.0
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
Faker==13.4.0
filelock==3.6.0
huggingface-hub==0.5.1
idna==3.3
joblib==1.1.0
mysqlclient==2.1.0
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.18.0
typing_extensions==4.2.0
urllib3==1.26.9
